+++ packages/std/vitest.config.ts
@@ -1,19 +1,18 @@
-import { defineConfig } from "vitest/config";
+import publishable from "@prodkit/shared/vitest/publishable";
+import { defineConfig, mergeConfig } from "vitest/config";
 
-export default defineConfig({
-  test: {
-    passWithNoTests: true,
-    coverage: {
-      all: true,
-      include: ["src/**/*.ts"],
-      exclude: ["src/**/*.test.ts", "src/index.ts"],
-      reporter: ["text", "json", "html", "lcov"],
-      reportsDirectory: "coverage",
-    },
-    include: ["src/**/*.test.ts"],
-    typecheck: {
-      enabled: true,
+export default mergeConfig(
+  publishable,
+  defineConfig({
+    test: {
+      passWithNoTests: true,
+      coverage: {
+        exclude: ["src/**/*.test.ts", "src/index.ts"],
+      },
       include: ["src/**/*.test.ts"],
+      typecheck: {
+        include: ["src/**/*.test.ts"],
+      },
     },
-  },
-});
+  }),
+);